Add missing license file: COPYING.LGPL-2
REVIEW:123528

The current license file included is a copy of the GNU Lesser General Public
License version 2.1, while we are distributing a number of files that are
licensed under the GNU Library General Public License version 2 *only*, and
as such it's not really distributable by Debian. Adding the a copy of the GNU
Library General Public License version 2 solves the issue.
Point to the right license file.
These files were licensed under the LGPLv2 only, and pointing to a LGPLv2.1
license file.
This patch was created as after ktextwidgets was rejected by the Debian ftpmasters.
https://lists.debian.org/debian-qt-kde/2015/04/msg00255.html