API docs: partial revert of earlier commit, didn't really work.
Keeping the find dialog open leads to all sorts of problems,
the counting of matches is wrong, the position to start from is
ambiguous (if we change the pattern -- or options? -- then it's a new
search so we should start from beginning or from cursor again, and
otherwise it's more of a find next ... that requires some magic in the
lambda to guess what the user expected). Closing the find dialog is just
easier, find new (from beginning or cursor) != find next (from last
match).