Fix compilation for users of kregexpeditorinterface.h
The deprecation macros need to include the export header.

This was found while compiling kdepim-addons, due to
kdepim-addons/sieveeditor/regexeditorlineedit/regexpeditorlineedit.cpp
Well, it's still broken because it disables all deprecated API,
so it needs to be ported, but at least this file is now correct.

CCMAIL: nicolas.fella@gmx.de

NO_CHANGELOG