[KRichTextEdit] Fix scroll jumping around when horizontal rule is added
Summary:
Due to Qt bug 83605, it's not a good idea to `setTextCursor` while the cursor
is inside `beginEditBlock / endEditBlock` (scrollbar might jump to the top),
see KDE bug 195828. It's sufficient to move `setTextCursor` outside to fix it.
(it's never too late... :)

This patch also adds a test for this case (which currently fails due to
regression, see {D28819} and {D28964} for the fix).

BUG: 195828

Test Plan:
 1) Apply {D28964}
 2) `ctest`

Reviewers: #frameworks, dfaure

Reviewed By: dfaure

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D28966