Port QRegExp to QRegularExpression (take 2)
KReplaceDialogPrivate::_k_slotOk() doesn't seem to be used/connected to
anything as far as I can see, I commented it out and everything build
just fine. Anyway, it's now been ported to QRegularExpression (but not
tested).

Also wrap "#include <QRegExp>"... etc with deprecation wrappers, this
was forgotten when the code was ported to QRegularExpression the first
time around.