Handle RTL text selection the same way as Qt
Ctrl+del was deleting the word on the right regardless of text direction. Changed to be logical-based.

Ctrl+left, Ctrl+right was moving the cursor based on the logical direction, which doesn't really make since since you are saying "left", "right". Changed to be visual-based.

The logic now is the same as Qt.

BUG: 397922